pyproject.toml
[project]
name = "projectwise-flaskui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask[async]>=3.1.1",
    "mcp[cli]>=1.12.2",
    "mem0ai>=0.1.115",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.1.1",
    "tiktoken>=0.9.0",
]
